requirements.txt
asgiref==3.5.1
backports.zoneinfo==0.2.1
Django==4.0.4
pkg_resources==0.0.0
sqlparse==0.4.2
django-cpf==0.1.0
django-birthday==0.1.4
